# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.6.0
  description: |
    Seshat API — Plataforma de gestão educacional com registro fotográfico, equipes, turmas e ocorrências.

    ## Banco de dados

    O PostgreSQL 16 roda **dentro do container de desenvolvimento** e é gerenciado
    automaticamente pelo `entrypoint.sh`:

    1. **Inicialização do cluster** — `pg_createcluster` é executado somente se o
       volume estiver vazio (`PG_VERSION` ausente em `PGDATA`).
    2. **Persistência** — dados armazenados no volume Docker `carometro-pg-dev`
       montado em `/var/lib/postgresql/16/main`.
    3. **Schema** — `drizzle-kit push` roda automaticamente **somente quando** o
       hash dos arquivos `lib/db/src/**/*.ts` muda. O hash é persistido em
       `$PGDATA/.carometro_schema_hash`.
    4. **Seed do administrador** — executado **apenas na primeira inicialização**
       via marcador `$PGDATA/.carometro_initialized`. O script `seed-admin.ts` é
       idempotente (não duplica o usuário se já existir).

    ### Comandos úteis

    ```
    docker compose run --rm dev db:push       # forçar reaplicação do schema
    docker compose run --rm dev db:seed       # verificar/recriar administrador
    docker compose run --rm dev db:hash-reset # remover marcadores e forçar re-setup
    docker compose down -v && docker compose up --build  # reinício limpo
    ```
servers:
  - url: /api
...
paths:
...
  /usuarios:
    get:
      operationId: listUsuarios
      tags: [usuarios]
      summary: List all usuarios
      responses:
        "200":
          description: List of usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      operationId: createUsuario
      tags: [usuarios]
      summary: Create a usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioCreated"
...
  /usuarios/{id}/foto:
    put:
      operationId: uploadFotoUsuario
      tags: [usuarios]
      summary: Upload or update profile photo for a usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadFotoUsuarioBody"
      responses:
        "200":
          description: Updated usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "404":
          description: Usuario not found
...
components:
  schemas:
...
    UsuarioDisciplina:
      type: object
      required: [ofertaId, disciplinaId, disciplinaNome, cursoId, cursoNome, turnoId, turnoNome]
      properties:
        ofertaId:
          type: string
          format: uuid
        disciplinaId:
          type: string
          format: uuid
        disciplinaNome:
          type: string
        cursoId:
          type: string
          format: uuid
        cursoNome:
          type: string
        turnoId:
          type: string
          format: uuid
        turnoNome:
          type: string
...
    UploadFotoUsuarioBody:
      type: object
      required: [fotoBase64]
      properties:
        fotoBase64:
          type: string
          description: Base64-encoded JPEG image (data URL)
    Usuario:
      type: object
      required: [id, email, codigoAcesso, primeiroAcesso, fotoUrl, roles, disciplinas, criadoEm]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
          nullable: true
        email:
          type: string
        dataNascimento:
          type: string
          format: date
          nullable: true
          description: Data de nascimento (YYYY-MM-DD). Obrigatória quando o perfil "estudante" for atribuído a menores de 18 anos.
        codigoAcesso:
          type: string
        primeiroAcesso:
          type: boolean
        fotoUrl:
          type: string
          nullable: true
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleSummary"
          description: >
            Um usuário pode ter múltiplos perfis simultaneamente.
            Exceção: usuários com o perfil "estudante" e menores de 18 anos
            só podem ter esse único perfil — nenhum outro pode ser combinado.
        disciplinas:
          type: array
          items:
            $ref: "#/components/schemas/UsuarioDisciplina"
        criadoEm:
          type: string
    RoleSummary:
      type: object
      required: [id, nome]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
    UsuarioCreated:
      type: object
      required: [usuario, senhaGerada]
      properties:
        usuario:
          $ref: "#/components/schemas/Usuario"
        senhaGerada:
          type: string
          description: >
            Plain text password shown only once. An e-mail with these
            credentials is also sent to the user's address automatically.
    UsuarioInput:
      type: object
      required: [email]
      properties:
        nome:
          type: string
          nullable: true
        email:
          type: string
        dataNascimento:
          type: string
          format: date
          nullable: true
          description: YYYY-MM-DD. Obrigatória se roleIds incluir o perfil "estudante" e o usuário for menor de 18 anos.
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          description: >
            IDs dos perfis a atribuir. Múltiplos perfis são permitidos.
            Restrição: se o perfil "estudante" for incluído e dataNascimento
            indicar idade < 18 anos, nenhum outro perfil pode ser combinado.
        disciplinaOfertaIds:
          type: array
          items:
            type: string
            format: uuid
...